Quorum support for TransactionalMap and TransactionalSet
