+++ hazelcast-client/src/test/java/com/hazelcast/client/quorum/list/ClientListQuorumReadTest.java
@@ -1,0 +1,41 @@
+package com.hazelcast.client.quorum.list;
+
+import com.hazelcast.client.quorum.PartitionedClusterClients;
+import com.hazelcast.client.test.TestHazelcastFactory;
+import com.hazelcast.config.Config;
+import com.hazelcast.core.IList;
+import com.hazelcast.quorum.list.ListQuorumReadTest;
+import com.hazelcast.test.HazelcastParametersRunnerFactory;
+import com.hazelcast.test.annotation.QuickTest;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.experimental.categories.Category;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+@RunWith(Parameterized.class)
+@Parameterized.UseParametersRunnerFactory(HazelcastParametersRunnerFactory.class)
+@Category({QuickTest.class})
+public class ClientListQuorumReadTest extends ListQuorumReadTest {
+
+    private static PartitionedClusterClients clients;
+
+    @BeforeClass
+    public static void setUp() {
+        TestHazelcastFactory factory = new TestHazelcastFactory();
+        initTestEnvironment(new Config(), factory);
+        clients = new PartitionedClusterClients(cluster, factory);
+    }
+
+    @AfterClass
+    public static void tearDown() {
+        shutdownTestEnvironment();
+        clients.terminateAll();
+    }
+
+    @Override
+    protected IList list(int index) {
+        return clients.client(index).getList(LIST_NAME + quorumType.name());
+    }
+
+}
