+++ hazelcast-client/src/test/java/com/hazelcast/client/quorum/list/ClientListQuorumWriteTest.java
@@ -1,11 +1,12 @@
-package com.hazelcast.client.quorum.set;
+package com.hazelcast.client.quorum.list;
 
 import com.hazelcast.client.quorum.PartitionedClusterClients;
 import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.config.Config;
-import com.hazelcast.core.ISet;
-import com.hazelcast.quorum.set.SetReadQuorumTest;
+import com.hazelcast.core.IList;
+import com.hazelcast.quorum.list.ListQuorumWriteTest;
 import com.hazelcast.test.HazelcastParametersRunnerFactory;
+import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -15,26 +16,28 @@ import org.junit.runners.Parameterized;
 
 @RunWith(Parameterized.class)
 @Parameterized.UseParametersRunnerFactory(HazelcastParametersRunnerFactory.class)
-@Category({QuickTest.class})
-public class ClientSetReadQuorumTest extends SetReadQuorumTest {
+@Category({QuickTest.class, ParallelTest.class})
+public class ClientListQuorumWriteTest extends ListQuorumWriteTest {
 
-    private static PartitionedClusterClients CLIENTS;
+    private static PartitionedClusterClients clients;
 
     @BeforeClass
     public static void setUp() {
         TestHazelcastFactory factory = new TestHazelcastFactory();
         initTestEnvironment(new Config(), factory);
-        CLIENTS = new PartitionedClusterClients(CLUSTER, factory);
+        clients = new PartitionedClusterClients(cluster, factory);
     }
 
     @AfterClass
     public static void tearDown() {
         shutdownTestEnvironment();
-        CLIENTS.terminateAll();
+        clients.terminateAll();
     }
 
-    protected ISet set(int index) {
-        return CLIENTS.client(index).getSet(SET_NAME + quorumType.name());
+    @Override
+    protected IList list(int index) {
+        return clients.client(index).getList(LIST_NAME + quorumType.name());
     }
+
 
 }
