+++ hazelcast-client/src/test/java/com/hazelcast/client/quorum/list/ClientTransactionalListQuorumReadTest.java
@@ -3,10 +3,10 @@ package com.hazelcast.client.quorum.list;
 import com.hazelcast.client.quorum.PartitionedClusterClients;
 import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.config.Config;
-import com.hazelcast.core.IList;
-import com.hazelcast.quorum.list.ListReadQuorumTest;
+import com.hazelcast.quorum.list.TransactionalListQuorumReadTest;
 import com.hazelcast.test.HazelcastParametersRunnerFactory;
 import com.hazelcast.test.annotation.QuickTest;
+import com.hazelcast.transaction.TransactionContext;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
@@ -16,25 +16,26 @@ import org.junit.runners.Parameterized;
 @RunWith(Parameterized.class)
 @Parameterized.UseParametersRunnerFactory(HazelcastParametersRunnerFactory.class)
 @Category({QuickTest.class})
-public class ClientListReadQuorumTest extends ListReadQuorumTest {
+public class ClientTransactionalListQuorumReadTest extends TransactionalListQuorumReadTest {
 
-    private static PartitionedClusterClients CLIENTS;
+    private static PartitionedClusterClients clients;
 
     @BeforeClass
     public static void setUp() {
         TestHazelcastFactory factory = new TestHazelcastFactory();
         initTestEnvironment(new Config(), factory);
-        CLIENTS = new PartitionedClusterClients(CLUSTER, factory);
+        clients = new PartitionedClusterClients(cluster, factory);
     }
 
     @AfterClass
     public static void tearDown() {
         shutdownTestEnvironment();
-        CLIENTS.terminateAll();
+        clients.terminateAll();
     }
 
-    protected IList list(int index) {
-        return CLIENTS.client(index).getList(LIST_NAME + quorumType.name());
+    @Override
+    public TransactionContext newTransactionContext(int index) {
+        return clients.client(index).newTransactionContext(options);
     }
 
 }
