+++ hazelcast-client/src/test/java/com/hazelcast/client/quorum/set/ClientSetQuorumWriteTest.java
@@ -1,12 +1,11 @@
-package com.hazelcast.client.quorum.list;
+package com.hazelcast.client.quorum.set;
 
 import com.hazelcast.client.quorum.PartitionedClusterClients;
 import com.hazelcast.client.test.TestHazelcastFactory;
 import com.hazelcast.config.Config;
-import com.hazelcast.core.IList;
-import com.hazelcast.quorum.list.ListWriteQuorumTest;
+import com.hazelcast.core.ISet;
+import com.hazelcast.quorum.set.SetQuorumWriteTest;
 import com.hazelcast.test.HazelcastParametersRunnerFactory;
-import com.hazelcast.test.annotation.ParallelTest;
 import com.hazelcast.test.annotation.QuickTest;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -16,27 +15,27 @@ import org.junit.runners.Parameterized;
 
 @RunWith(Parameterized.class)
 @Parameterized.UseParametersRunnerFactory(HazelcastParametersRunnerFactory.class)
-@Category({QuickTest.class, ParallelTest.class})
-public class ClientListWriteQuorumTest extends ListWriteQuorumTest {
+@Category({QuickTest.class})
+public class ClientSetQuorumWriteTest extends SetQuorumWriteTest {
 
-    private static PartitionedClusterClients CLIENTS;
+    private static PartitionedClusterClients clients;
 
     @BeforeClass
     public static void setUp() {
         TestHazelcastFactory factory = new TestHazelcastFactory();
         initTestEnvironment(new Config(), factory);
-        CLIENTS = new PartitionedClusterClients(CLUSTER, factory);
+        clients = new PartitionedClusterClients(cluster, factory);
     }
 
     @AfterClass
     public static void tearDown() {
         shutdownTestEnvironment();
-        CLIENTS.terminateAll();
+        clients.terminateAll();
     }
 
-    protected IList list(int index) {
-        return CLIENTS.client(index).getList(LIST_NAME + quorumType.name());
+    @Override
+    protected ISet set(int index) {
+        return clients.client(index).getSet(SET_NAME + quorumType.name());
     }
-
 
 }
