+++ hazelcast/src/main/java/com/hazelcast/collection/impl/txncollection/operations/CollectionTxnAddBackupOperation.java
@@ -23,10 +23,11 @@ import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.spi.BackupOperation;
+import com.hazelcast.spi.impl.MutatingOperation;
 
 import java.io.IOException;
 
-public class CollectionTxnAddBackupOperation extends CollectionOperation implements BackupOperation {
+public class CollectionTxnAddBackupOperation extends CollectionOperation implements BackupOperation, MutatingOperation {
 
     private long itemId;
     private Data value;
