+++ hazelcast/src/main/java/com/hazelcast/collection/impl/txncollection/operations/CollectionTxnAddOperation.java
@@ -25,10 +25,12 @@ import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.spi.Operation;
+import com.hazelcast.spi.impl.MutatingOperation;
 
 import java.io.IOException;
 
-public class CollectionTxnAddOperation extends CollectionBackupAwareOperation implements CollectionTxnOperation {
+public class CollectionTxnAddOperation extends CollectionBackupAwareOperation implements CollectionTxnOperation,
+        MutatingOperation {
 
     private long itemId;
     private Data value;
