+++ hazelcast/src/main/java/com/hazelcast/collection/impl/txncollection/operations/CollectionTxnRemoveBackupOperation.java
@@ -22,10 +22,11 @@ import com.hazelcast.collection.impl.collection.operations.CollectionOperation;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.spi.BackupOperation;
+import com.hazelcast.spi.impl.MutatingOperation;
 
 import java.io.IOException;
 
-public class CollectionTxnRemoveBackupOperation extends CollectionOperation implements BackupOperation {
+public class CollectionTxnRemoveBackupOperation extends CollectionOperation implements BackupOperation, MutatingOperation {
 
     private long itemId;
