+++ hazelcast/src/main/java/com/hazelcast/collection/impl/txncollection/operations/CollectionTxnRemoveOperation.java
@@ -25,10 +25,12 @@ import com.hazelcast.core.ItemEventType;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.ObjectDataOutput;
 import com.hazelcast.spi.Operation;
+import com.hazelcast.spi.impl.MutatingOperation;
 
 import java.io.IOException;
 
-public class CollectionTxnRemoveOperation extends CollectionBackupAwareOperation implements CollectionTxnOperation {
+public class CollectionTxnRemoveOperation extends CollectionBackupAwareOperation implements CollectionTxnOperation,
+        MutatingOperation {
 
     private long itemId;
     private transient CollectionItem item;
