+++ hazelcast/src/test/java/com/hazelcast/quorum/set/AbstractSetQuorumTest.java
@@ -35,7 +35,7 @@ public abstract class AbstractSetQuorumTest {
 
     protected static final String SET_NAME = "quorum" + randomString();
 
-    protected static PartitionedCluster CLUSTER;
+    protected static PartitionedCluster cluster;
 
     protected static void initTestEnvironment(Config config, TestHazelcastInstanceFactory factory) {
         initCluster(PartitionedCluster.createClusterConfig(config), factory, READ, WRITE, READ_WRITE);
@@ -43,7 +43,7 @@ public abstract class AbstractSetQuorumTest {
 
     protected static void shutdownTestEnvironment() {
         HazelcastInstanceFactory.terminateAll();
-        CLUSTER = null;
+        cluster = null;
     }
 
 
@@ -63,7 +63,7 @@ public abstract class AbstractSetQuorumTest {
     }
 
     protected static void initCluster(Config config, TestHazelcastInstanceFactory factory, QuorumType... types) {
-        CLUSTER = new PartitionedCluster(factory);
+        cluster = new PartitionedCluster(factory);
 
         String[] quorumNames = new String[types.length];
         int i = 0;
@@ -76,12 +76,12 @@ public abstract class AbstractSetQuorumTest {
             quorumNames[i++] = quorumName;
         }
 
-        CLUSTER.createFiveMemberCluster(config);
-        CLUSTER.splitFiveMembersThreeAndTwo(quorumNames);
+        cluster.createFiveMemberCluster(config);
+        cluster.splitFiveMembersThreeAndTwo(quorumNames);
     }
 
     protected ISet set(int index, QuorumType quorumType) {
-        return CLUSTER.instance[index].getSet(SET_NAME + quorumType.name());
+        return cluster.instance[index].getSet(SET_NAME + quorumType.name());
     }
 
 }
